Implement suggestions from code review
