--- gingado/temporal/__init__.py
@@ -1,1 +1,0 @@
-from gingado.temporal.features import get_timefeat, validate_and_get_freq
