+++ tests/temporal/test_features.py
@@ -5,8 +5,8 @@ import pandas as pd
 import pytest
 from pandas.testing import assert_frame_equal
 
-import gingado.temporal.features as tp
-from gingado.temporal.types import (
+from gingado.utils import get_timefeat, dates_Xy
+from gingado.internals import (
     DayFeatures,
     Frequency,
     InvalidTemporalFeature,
@@ -14,6 +14,8 @@ from gingado.temporal.types import (
     QuarterFeatures,
     WeekFeatures,
 )
+
+import gingado.internals as tp
 
 
 @pytest.fixture(name="sample_df")
@@ -103,7 +105,7 @@ def test_get_timefeat_add_to_df(
     sample_df: pd.DataFrame, frequency: Frequency, num_columns: int
 ) -> None:
     """Test get_timefeat function when add_to_df is True."""
-    result = tp.get_timefeat(sample_df, freq=frequency, add_to_df=True)
+    result = get_timefeat(sample_df, freq=frequency, add_to_df=True)
     assert (
         len(result.columns) == len(sample_df.columns) + num_columns
     )  # Original column + num_columns time features
@@ -124,7 +126,7 @@ def test_get_timefeat_return_only_features(
     sample_df: pd.DataFrame, frequency: Frequency, num_columns
 ) -> None:
     """Test get_timefeat function when add_to_df is False."""
-    result = tp.get_timefeat(sample_df, freq=frequency, add_to_df=False)
+    result = get_timefeat(sample_df, freq=frequency, add_to_df=False)
     assert len(result.columns) == num_columns  # Only time features
     assert len(result) == len(sample_df)  # Same number of rows as before
     assert "value" not in result.columns
@@ -135,12 +137,12 @@ def test_get_timefeat_invalid_index(frequency: Frequency) -> None:
     """Test get_timefeat function with an invalid DataFrame index."""
     invalid_df = pd.DataFrame({"value": range(10)})
     with pytest.raises(ValueError):
-        tp.get_timefeat(invalid_df, freq=frequency)
+        get_timefeat(invalid_df, freq=frequency)
 
 
 def test_feature_values(sample_df: pd.DataFrame) -> None:
     """Test specific feature values generated by get_timefeat"""
-    result = tp.get_timefeat(sample_df, freq=Frequency.DAILY, add_to_df=False)
+    result = get_timefeat(sample_df, freq=Frequency.DAILY, add_to_df=False)
     assert result.loc["2023-01-01", "day_of_week"] == 6  # Sunday
     assert result.loc["2023-01-01", "day_of_month"] == 1
     assert result.loc["2023-01-01", "day_of_year"] == 1
@@ -152,7 +154,7 @@ def test_feature_values(sample_df: pd.DataFrame) -> None:
 
 def test_user_provided_features(sample_df: pd.DataFrame) -> None:
     """Tests that only user provided temporal feature are used"""
-    result = tp.get_timefeat(
+    result = get_timefeat(
         sample_df, freq=Frequency.DAILY, columns=["day_of_week", "month_of_year"], add_to_df=False
     )
 
@@ -167,7 +169,7 @@ def test_user_provided_wrong_frequency_features(monthly_start_df: pd.DataFrame) 
         UserWarning,
         match="Requested temporal feature day_of_week not available for data with frequency MS!",
     ):
-        result = tp.get_timefeat(
+        result = get_timefeat(
             monthly_start_df,
             freq=Frequency.MONTHLY,
             columns=["day_of_week", "month_of_year"],
@@ -182,7 +184,7 @@ def test_user_provided_invalid_frequency(sample_df: pd.DataFrame) -> None:
     with pytest.raises(
         InvalidTemporalFeature, match="Invalid temporal feature passed: 'ace_of_spades'"
     ):
-        tp.get_timefeat(sample_df, freq=Frequency.DAILY, columns=["day_of_year", "ace_of_spades"])
+        get_timefeat(sample_df, freq=Frequency.DAILY, columns=["day_of_year", "ace_of_spades"])
 
 
 def test_day_features(sample_df: pd.DataFrame) -> None:
@@ -331,7 +333,7 @@ def test_dates_Xy_basic(sample_data):
     """Tests the basic structure returned by dates_Xy"""
     X, y = sample_data
     date_list = [datetime(2023, 3, 1), datetime(2023, 6, 1), datetime(2023, 9, 1)]
-    result = tp.dates_Xy(X, y, date_list)
+    result = dates_Xy(X, y, date_list)
 
     #
     assert isinstance(result, list)
@@ -350,7 +352,7 @@ def test_dates_Xy_filtered_data(sample_data):
     """Tests that filtering works as expected"""
     X, y = sample_data
     date_list = [datetime(2023, 3, 1), datetime(2023, 6, 1)]
-    result = tp.dates_Xy(X, y, date_list)
+    result = dates_Xy(X, y, date_list)
 
     assert len(result[0][0]["d"]) == 60  # 31 + 28 + 1 days
     assert len(result[1][0]["d"]) == 152  # 151 days from Jan to May and 1st of June
@@ -365,7 +367,7 @@ def test_dates_Xy_empty_X(sample_data):
         index=pd.date_range(start="2023-03-01", end="2023-03-01", freq="D"),
         data={"month_of_quarter": 3, "month_of_year": 3},
     )
-    result = tp.dates_Xy(X, y, date_list)
+    result = dates_Xy(X, y, date_list)
 
     assert len(result) == 1
     assert list(result[0][0].keys()) == ["future"]
@@ -376,7 +378,7 @@ def test_dates_Xy_empty_date_list(sample_data):
     """Tests the dates_Xy method when an empty date list is passed"""
     X, y = sample_data
     date_list = []
-    result = tp.dates_Xy(X, y, date_list)
+    result = dates_Xy(X, y, date_list)
 
     assert not result
 
@@ -388,7 +390,7 @@ def test_dates_Xy_different_frequencies(sample_data):
         index=pd.date_range(start="2023-01-01", end="2023-12-31", freq="MS"),
     )
     date_list = [datetime(2023, 6, 15)]
-    result = tp.dates_Xy(X, y, date_list)
+    result = dates_Xy(X, y, date_list)
 
     assert set(result[0][0].keys()) == {"d", "w", "m", "future"}
     assert len(result[0][0]["d"]) == 166  # Days from Jan 1 to Jun 15
@@ -399,7 +401,7 @@ def test_dates_Xy_different_frequencies(sample_data):
 def test_dates_Xy_custom_freq_y(sample_data):
     X, y = sample_data
     date_list = [datetime(2023, 6, 1)]
-    result = tp.dates_Xy(X, y, date_list, freq_y="W")
+    result = dates_Xy(X, y, date_list, freq_y="W")
 
     future_features = result[0][0]["future"]
     assert "week_of_month" in future_features.columns
